requirements.txt
django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
dj-database-url==0.5.0
gunicorn==20.1.0
whitenoise==5.3.0
psycopg2==2.9.1
memory-profiler==0.58.0
